server/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
pydantic-settings
dnspython
python-dotenv
pandas
numpy
scikit-learn
groq
bcrypt
google-auth
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
requests
